requirements.txt
# Core transcription
whisperx>=3.1.1
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
pydub>=0.25.1
soundfile>=0.12.1

# Speaker diarization (pulled by whisperx, but explicit)
pyannote.audio>=3.1.1

# Utilities
click>=8.1.0
pyyaml>=6.0
rich>=13.0.0
python-dotenv>=1.0.0

# For recap generation (optional - can use external LLM)
openai>=1.0.0
requests>=2.31.0

# Web interface
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sse-starlette>=1.8.0
jinja2>=3.1.0
aiofiles>=23.0.0
markdown>=3.5.0
